Revert "Modularize code"
