--- .devcontainer/devcontainer.json
@@ -1,6 +1,0 @@
-{
-  "tasks": {
-    "build": "pip install -r requirements.txt",
-    "launch": "python -m deeptuner.run"
-  }
-}
