--- deeptune/losses/triplet_loss.py
@@ -1,11 +1,0 @@
-import tensorflow as tf
-
-def triplet_loss(margin=1.0):
-    def loss(y_true, y_pred):
-        anchor, positive, negative = y_pred[0], y_pred[1], y_pred[2]
-        pos_dist = tf.reduce_sum(tf.square(anchor - positive), axis=-1)
-        neg_dist = tf.reduce_sum(tf.square(anchor - negative), axis=-1)
-        basic_loss = pos_dist - neg_dist + margin
-        loss = tf.reduce_mean(tf.maximum(basic_loss, 0.0), axis=0)
-        return loss
-    return loss
